[BE] FEAT: 테스트할 파일 생성 및 ApiRequestManager 유닛 테스트 작성
