+++ backend/src/test/java/org/ftclub/cabinet/auth/domain/AuthAspectUnitTest.java
@@ -1,0 +1,5 @@
+package org.ftclub.cabinet.auth.domain;
+
+class AuthAspectUnitTest {
+
+}
