+++ backend/src/test/java/org/ftclub/cabinet/auth/service/AuthFacadeServiceImplUnitTest.java
@@ -1,0 +1,22 @@
+package org.ftclub.cabinet.auth.service;
+
+import org.junit.jupiter.api.Test;
+
+class AuthFacadeServiceImplUnitTest {
+
+    @Test
+    void requestLoginToApi() {
+    }
+
+    @Test
+    void handleLogin() {
+    }
+
+    @Test
+    void masterLogin() {
+    }
+
+    @Test
+    void logout() {
+    }
+}
