+++ backend/src/test/java/org/ftclub/cabinet/auth/service/AuthServiceUnitTest.java
@@ -1,0 +1,14 @@
+package org.ftclub.cabinet.auth.service;
+
+import org.junit.jupiter.api.Test;
+
+class AuthServiceUnitTest {
+
+    @Test
+    void validateMasterLogin() {
+    }
+
+    @Test
+    void addUserIfNotExistsByClaims() {
+    }
+}
